+++ CHANGELOG.md
@@ -10,6 +10,7 @@ v2.7.3
   constructor.
 * Support logical shifts by longs, although only the lower 5 or 6 bits are effective.
 * Fixed an infinite loop when looking for superclasses of java.lang.Object.
+* Improve peephole optimization against the relational operations which return a boolean.
 
 v2.7.2 (2026-02-23)
 ------